% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFns.R
\name{checkSpelling}
\alias{checkSpelling}
\title{check for spelling errors in a character vector}
\usage{
checkSpelling(input)
}
\arguments{
\item{input}{character vector}
}
\description{
takes character and returns list of problematic
    spellings and suggestions for replacement
}
