% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTemplate.R
\name{checkTemplate}
\alias{checkTemplate}
\title{Check template data frame}
\usage{
checkTemplate(df, ImmPortTemplateName)
}
\arguments{
\item{df}{data frame with user entered data}

\item{ImmPortTemplateName}{Name of ImmPort Template to compare against}
}
\description{
Check template data frame based on class, dimensions, colnames,
    data types, and format.
}
