% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoData}
\alias{demoData}
\title{demoData}
\format{A list of data frames holding demo data for the basic study design
    template.}
\usage{
demoData
}
\description{
demoData
}
\keyword{datasets}
