% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFns.R
\name{findReplace}
\alias{findReplace}
\title{find and replace characters in a vector}
\usage{
findReplace(input, find, replace)
}
\arguments{
\item{input}{character vector}

\item{find}{word to find}

\item{replace}{replacement word}
}
\description{
find all instances of a word in a vector
    and replace it with another word
}
