% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2immport.R
\name{geo2ImmPort}
\alias{geo2ImmPort}
\title{Output ImmPort-compliant tsv files for gene expression matrices from GEO}
\usage{
geo2ImmPort(geoGSE, outputFile = NULL)
}
\arguments{
\item{geoGSE}{string, GEO Series accession}

\item{outputFile}{string, filepath for output. default is NULL.}
}
\description{
Given a GEO Series accession, geo2ImmPort will output a
}
