% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookupFns.R
\name{getLookups}
\alias{getLookups}
\title{retrieve colnames for an ImmPort Template with associated lookup tables}
\usage{
getLookups(ImmPortTemplateName)
}
\arguments{
\item{ImmPortTemplateName}{name of the ImmPort Template to check}
}
\description{
Some ImmPort Templates have controlled or preferred values
    for certain columns that are found in 'lookup' tables.  This function
    returns the colnames that have associated lookup tables.
}
