% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookupFns.R
\name{getLookupValues}
\alias{getLookupValues}
\title{retrieve allowed or preferred values for a given template and column}
\usage{
getLookupValues(ImmPortTemplateName, templateColname)
}
\arguments{
\item{ImmPortTemplateName}{ImmPort Template Name}

\item{templateColname}{column name in template}
}
\description{
Given a template and column, this function returns the values
   that are allowed or preferred for this entry.  The return object is a
   vector to make searching easier.
}
