% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImmPortLookups}
\alias{ImmPortLookups}
\title{ImmPortLookups}
\format{A data frame holding ImmPort lookup information from lookups.json
    that is used to create and check empty templates for user editing.}
\usage{
ImmPortLookups
}
\description{
ImmPortLookups
}
\keyword{datasets}
