% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImmPortTemplates}
\alias{ImmPortTemplates}
\title{ImmPortTemplates}
\format{
A data frame holding ImmPort Schema information from templates.json
    that is used to create and check empty templates for user editing.
}
\usage{
ImmPortTemplates
}
\description{
ImmPortTemplates
}
\keyword{datasets}
