% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFns.R
\name{interactiveReplace}
\alias{interactiveReplace}
\title{Interactively find and replace problematic words in a vector}
\usage{
interactiveReplace(checkSpellingOutput, inputVector)
}
\arguments{
\item{checkSpellingOutput}{named list of problematic words and suggested replacements}

\item{inputVector}{character vector}
}
\description{
For each word in the output list from checkSpelling()
    find the word in the inputVector and allow the user to enter a
    replacement word
}
