% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFns.R
\name{interactiveSpellCheck.df}
\alias{interactiveSpellCheck.df}
\title{Interactively check spelling against dictionary and ImmuneSpace-specific words}
\usage{
interactiveSpellCheck.df(inputDF, outputDir, dfName = NULL)
}
\arguments{
\item{inputDF}{dataframe, only character type will be worked on}

\item{outputDir}{filepath for where to append lines of code}

\item{dfName}{name of df for use in output R doc, default NULL uses "templateName" attribute}
}
\description{
Given an input dataframe and output directory, the user may
    correct words that are not found in a standard dictionary or a
    current list of ImmuneSpace specific terms.
}
