% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISFreqsAll}
\alias{ISFreqsAll}
\title{ISFreqsAll}
\format{
A named list of word frequencies in certain table and column name
    combinations of ImmuneSpace
}
\usage{
ISFreqsAll
}
\description{
ISFreqsAll
}
\keyword{datasets}
