% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISFreqsByNm}
\alias{ISFreqsByNm}
\title{ISFreqsByNm}
\format{A named list of lists of word frequencies in certain table and column
    combinations of ImmuneSpace. Higher level names are paste0(table, '-', column).}
\usage{
ISFreqsByNm
}
\description{
ISFreqsByNm
}
\keyword{datasets}
