% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBSD.R
\name{parseBSD}
\alias{parseBSD}
\title{parse the basic study design template and return a list of R data frames}
\usage{
parseBSD(filePath)
}
\arguments{
\item{filePath}{file path for basic study design csv. Cannot be excel spreadsheet.}
}
\description{
function parses a csv version of the basic study design template
    from ImmPort to generate a list of R data frames that can then be checked with
    the quality control functions in the package.
}
