% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{transform_Data}
\alias{transform_Data}
\title{Transform R dataframe into ImmPort csv format for non-Basic Study Design templates}
\usage{
transform_Data(df, outputDir = NULL, validate = TRUE)
}
\arguments{
\item{df}{dataframe holding non-header template information}

\item{outputDir}{filepath to directory for output csv}

\item{validate}{boolean determining whether to use ImmPort validation scripts on output csv}
}
\description{
Based on ImmPort's specifications, the function validates
a data frame holding meta or assay data and then outputs a csv
for submission to the database.
}
