% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_txt}
\alias{write_txt}
\title{Write a list of blocks to a text file}
\usage{
write_txt(name, blocks, file)
}
\arguments{
\item{name}{A character of the name.}

\item{blocks}{A list of blocks (lists or data.frames).}

\item{file}{A character of the file path.}
}
\value{
returns NULL invisibly.
}
\description{
Write a list of blocks to a text file
}
