% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTemplate.R
\name{checkTemplate}
\alias{checkTemplate}
\title{Check template data frame}
\usage{
checkTemplate(df)
}
\arguments{
\item{df}{data frame with user entered data and 'templateName' attribute}
}
\description{
Check template data frame based on class, dimensions, colnames,
    data types, and format.
}
