% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTemplate.R
\name{checkTemplate}
\alias{checkTemplate}
\title{Check template data frame}
\usage{
checkTemplate(df, chkTypes = TRUE, quiet = TRUE)
}
\arguments{
\item{df}{data frame with user entered data and 'templateName' attribute}

\item{chkTypes}{boolean, TRUE by default, checks types of data by column}

\item{quiet}{boolean, TRUE by default, instead of errors, returns problems}
}
\description{
Check template data frame based on class, dimensions, colnames,
    data types, and format.
}
