% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2immport.R
\name{geo2ImmPort}
\alias{geo2ImmPort}
\title{Output ImmPort-compliant tsv files for gene expression matrices from GEO}
\usage{
geo2ImmPort(softGzPath, outputDir = NULL, stdNorm = TRUE)
}
\arguments{
\item{softGzPath}{string, GEO .soft.gz local filepath}

\item{outputDir}{string, filepath for output. default is NULL.}

\item{stdNorm}{boolean, standardize and normalize exprMx.}
}
\description{
Given a GEO Series accession, geo2ImmPort will output a
}
