% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplateDF.R
\name{getTemplateDF}
\alias{getTemplateDF}
\title{generate a template data frame for a given ImmPort Template Name}
\usage{
getTemplateDF(ImmPortTemplateName, rowNum = 1)
}
\arguments{
\item{ImmPortTemplateName}{name of the ImmPort Template to create}

\item{rowNum}{sets the number of rows for the ImmPort Template. Defaults to one row.}
}
\description{
Based on ImmPort's specifications, the function creates an
    empty dataframe with the correct column headers for the user to update
    as part of the ImmPort submission development.
}
