% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTextFns.R
\name{interactiveSpellCheck.df}
\alias{interactiveSpellCheck.df}
\title{Interactively check spelling against dictionary and ImmuneSpace-specific words}
\usage{
interactiveSpellCheck.df(inputDF, vectorName, outputDir)
}
\arguments{
\item{inputDF}{dataframe, only character type will be worked on}

\item{vectorName}{name of vector for use in output R doc}

\item{outFile}{filepath for where to append lines of code}
}
\description{
Given an input vector and output directory, the user may
    correct words that are not found in a standard dictionary or a
    current list of ImmuneSpace specific terms.
}
