% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_basicStudyDesign.R
\name{transform_basicStudyDesign}
\alias{transform_basicStudyDesign}
\title{Transform R dataframes into ImmPort basicStudyDesign.csv format}
\usage{
transform_basicStudyDesign(study, arm_or_cohort, study_personnel, planned_visit,
  inclusion_exclusion, study_2_protocol, study_file, study_link, study_pubmed,
  outputDir = NULL, validate = TRUE)
}
\arguments{
\item{study}{a two-column dataframe with names and values for study metaData}

\item{arm_or_cohort}{four column df defining arms or cohorts in study}

\item{study_personnel}{ten column df identifying all study personnel}

\item{planned_visit}{seven column df describing all visits and potential variation}

\item{inclusion_exclusion}{three column df defining allowed participant criteria}

\item{study_2_protocol}{small 1 x 2 df with protocol reference}

\item{study_file}{three column df describing all related files being submitted}

\item{study_link}{two column df with web links for study}

\item{study_pubmed}{nine column df identifying publication information}

\item{outputDir}{filepath to directory for output csv}

\item{validate}{boolean determining whether to use ImmPort validation scripts on output csv}
}
\description{
Based on ImmPort's specifications, the function validates
a group of dataframes holding metaData and then outputs a csv
for submission to the database.
}
